+++ CHANGELOG.md
@@ -7,6 +7,14 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Added two new events: BeforeChunk and AfterChunk for chunked exports
 
+## [3.1.40] - 2022-05-02
+
+### Changed
+
+- Adds `WithDefaultStyles` concern to allow configuring the workbook default styles.
+- Adds `WithBackgroundColor` concern to allow configuring the workbook default background color.
+- Expose the ability to set custom response headers when exporting collections via Exportable
+
 ## [3.1.39] - 2022-04-23
 
 ### Fixed
@@ -226,7 +234,8 @@ All notable changes to this project will be documented in this file.
 - Raw() method now also available on Exportable.
 - Fix for breaking changes in PhpSpreadsheet with empty enclosures.
 
-[Unreleased]: https://github.com/Maatwebsite/Laravel-Excel/compare/3.1.39...HEAD
+[Unreleased]: https://github.com/Maatwebsite/Laravel-Excel/compare/3.1.40...HEAD
+[3.1.40]: https://github.com/Maatwebsite/Laravel-Excel/compare/3.1.39...3.1.40
 [3.1.39]: https://github.com/Maatwebsite/Laravel-Excel/compare/3.1.38...3.1.39
 [3.1.38]: https://github.com/Maatwebsite/Laravel-Excel/compare/3.1.37...3.1.38
 [3.1.37]: https://github.com/Maatwebsite/Laravel-Excel/compare/3.1.36...3.1.37